// ImageViewerPCF/ImageViewerPCF/generated/ManifestTypes.d.ts
/*
*This is auto generated from the ControlManifest.Input.xml file
*/

// Define IInputs and IOutputs Type. They should match with ControlManifest.
export interface IInputs {
    defectTempId: ComponentFramework.PropertyTypes.StringProperty;
    debugMode: ComponentFramework.PropertyTypes.TwoOptionsProperty;
}
export interface IOutputs {
    defectTempId?: string;
}